+++ src/App.js
@@ -1,16 +1,16 @@
 
 import './App.css';
-import Header from './component/Header'
 import Login from './component/Login'
 import SignUp from './component/SignUp'
-import {Route} from 'react-router-dom'
+import { Route } from 'react-router-dom'
+import { Home } from './component/home/Home';
 // comentario
 function App() {
   return (
     <>
-    <Header/>
-    <Route path="/login" component={Login} />
-    <Route path="/signup" render={() => <SignUp titulo='SignUp component 235'/>}/>
+      <Route exact path='/' component={Home} />
+      <Route path="/login" component={Login} />
+      <Route path="/signup" render={() => <SignUp titulo='SignUp component 235' />} />
     </>
   )
 }
