+++ src/component/home/Home.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export function Home() {
+    return (
+        <div>Hola</div>
+    )
+}
